+++ HISTORY.rst
@@ -12,6 +12,7 @@ History
 * Replaces TravisCI with Github Actions
 * Replaces requests with httpx
 * Replaces click with typer
+* Adds Pydantic PlayerState, Song and Episode models
 
 0.1.0 (2018-12-25)
 ------------------